src/katsdpnetboxutilities/system/requirements.txt
attrs==21.2.0
awesome-slugify==1.6.5
boltons==21.0.0
certifi==2021.5.30
chardet==4.0.0
ConfigArgParse==1.2.3
future==0.18.2
graphviz==0.14.1
idna==2.10
ipython-genutils==0.2.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
git+https://github.com/ska-sa/katsdpnetboxutilities.git
#katsdpnetboxutilities @ git+https://github.com/ska-sa/katsdpnetboxutilities.git@f1da95d256fa48f5c65c3e0c3d4cf5cd29f0e27d
katversion==1.1
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
nbformat==5.1.3
preprocess==2.0.0
publish-doconce==1.1.4
Pygments==2.9.0
pygments-doconce==1.0.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-Levenshtein-wheels==0.13.2
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
six==1.16.0
tornado==6.1
traitlets==5.0.5
unicode==2.8
Unidecode==0.4.21
urllib3==1.26.5
yml==0.0.1
